+++ lib/notification/model/campaign_app_data_dto.dart
@@ -1,7 +1,6 @@
 import 'dart:convert';
 
 import 'package:flutter/foundation.dart';
-import 'package:stoyco_shared/notification/model/data.dart';
 import 'package:stoyco_shared/notification/model/field.dart';
 
 @immutable
@@ -9,7 +8,7 @@ class CampaignAppDataDto {
   final String? userId;
   final String? itemId;
   final List<Field>? fields;
-  final Data? data;
+  final Map<String, String>? data;
 
   const CampaignAppDataDto({
     this.userId,
@@ -29,16 +28,15 @@ class CampaignAppDataDto {
         fields: (data['fields'] as List<dynamic>?)
             ?.map((e) => Field.fromMap(e as Map<String, dynamic>))
             .toList(),
-        data: data['data'] == null
-            ? null
-            : Data.fromMap(data['data'] as Map<String, dynamic>),
+        data: (data['data'] as Map<String, dynamic>?)
+            ?.map((key, value) => MapEntry(key, value as String)),
       );
 
   Map<String, dynamic> toMap() => {
         'userId': userId,
         'itemId': itemId,
         'fields': fields?.map((e) => e.toMap()).toList(),
-        'data': data?.toMap(),
+        'data': data,
       };
 
   /// `dart:convert`
@@ -56,7 +54,7 @@ class CampaignAppDataDto {
     String? userId,
     String? itemId,
     List<Field>? fields,
-    Data? data,
+    Map<String, String>? data,
   }) =>
       CampaignAppDataDto(
         userId: userId ?? this.userId,
@@ -68,4 +66,15 @@ class CampaignAppDataDto {
   @override
   int get hashCode =>
       userId.hashCode ^ itemId.hashCode ^ fields.hashCode ^ data.hashCode;
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+
+    return other is CampaignAppDataDto &&
+        other.userId == userId &&
+        other.itemId == itemId &&
+        listEquals(other.fields, fields) &&
+        mapEquals(other.data, data);
+  }
 }